Texture size is ivec3

+++ resources/shaders/glbspshader.frag
@@ -65,7 +65,7 @@ void main() {
     vec2 texcoords = vec2(0.0);
     vec2 texuvmod = vec2(0.0);
     vec2 deltas = vec2(0.0);
-    vec3 texsize = textureSize(textureArray0,0);
+    ivec3 texsize = textureSize(textureArray0,0);
 
     // texture flow mods
     if (abs(vsNorm.z) >= 0.9) {
@@ -114,10 +114,10 @@ void main() {
             curpos = curpos - constspeed * modetime + acceleration * modetime * modetime / 2.0;
         }
         deltas.x = 0.0;
-        deltas.y = mod(curpos, texsize.y);
+        deltas.y = mod(curpos, float(texsize.y));
     } else {
-        deltas.x = texuvmod.x * mod(float(flowtimer), texsize.x);
-        deltas.y = texuvmod.y * mod(float(flowtimer), texsize.y);
+        deltas.x = texuvmod.x * mod(float(flowtimer), float(texsize.x));
+        deltas.y = texuvmod.y * mod(float(flowtimer), float(texsize.y));
     }
 
     texcoords.x = (deltas.x + texuv.x) / float(texsize.x);
